Changed AddStateVariable() initial value argument to double type.
git-svn-id: http://www.bci2000.org/svn/trunk@3978 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/ExternalLinks/BCI2000Command/BCI2000Command.cpp
@@ -348,7 +348,7 @@ SetParameter( int, char** inArgs, BCI2000Remote& ioBCI )
 int
 AddStateVariable( int, char** inArgs, BCI2000Remote& ioBCI )
 {
-  return ioBCI.AddStateVariable( inArgs[0], ::atoi( inArgs[1] ), ::atoi( inArgs[2] ) ) ? ok : error;
+  return ioBCI.AddStateVariable( inArgs[0], ::atoi( inArgs[1] ), ::atof( inArgs[2] ) ) ? ok : error;
 }
 
 int
